:truck: Move data.lightning to data_loaders
