+++ ice_station_zebra/evaluation/callbacks/plotting_callback.py
@@ -6,7 +6,7 @@ from lightning.pytorch import Callback
 from torch import Tensor
 from torch.utils.data import DataLoader
 
-from ice_station_zebra.data.lightning import CombinedDataset
+from ice_station_zebra.data_loaders import CombinedDataset
 from ice_station_zebra.visualisations import plot_sic_comparison
 
 logger = logging.getLogger(__name__)
