+++ ice_station_zebra/evaluation/evaluator.py
@@ -5,7 +5,7 @@ import hydra
 from lightning import Callback, Trainer
 from omegaconf import DictConfig, OmegaConf
 
-from ice_station_zebra.data.lightning import ZebraDataModule
+from ice_station_zebra.data_loaders import ZebraDataModule
 from ice_station_zebra.models import ZebraModel
 from ice_station_zebra.utils import get_timestamp
