+++ ice_station_zebra/training/trainer.py
@@ -5,7 +5,7 @@ import hydra
 from lightning import Callback, Trainer
 from omegaconf import DictConfig, OmegaConf
 
-from ice_station_zebra.data.lightning import ZebraDataModule
+from ice_station_zebra.data_loaders import ZebraDataModule
 from ice_station_zebra.models import ZebraModel
 from ice_station_zebra.utils import generate_run_name, get_wandb_logger
